rhdeservice/scripts/regression/requirements.txt
pygal
jinja2
pygsheets
pandas
numpy
google-api-python-client